renaming file and adjusting contributor order
