+++ lib/feathers-nedb.js
@@ -1,6 +1,6 @@
 // Functions for modifying data / params
 var Proto = require('uberproto');
-var prepareSpecials = require('./filters/query.prepareSpecials');
+var prepareSpecials = require('./filters/prepare-specials');
 var Datastore = require('nedb');
 var errors = require('feathers-errors').types;
