+++ lib/filters/prepare-specials.js
